Move word_counter inside i in range so it resets
I suggest moving word_counter so it resets to 0 each time the for i in range runs.  I found your solution because I'm not understanding the object oriented programming aspect of this problem/ week at all.  You've nailed it except I think the word_counter needs to move inside the for loop so that it resets to 0 each time you try a different shift.  Apologies if I'm not using GitHub properly as this is literally my first interaction on the site.  Thanks, Scott